+++ rackcity/views/__init__.py
@@ -1,0 +1,1 @@
+from .it_model_views import model_detail, model_list
